Suppress testing on production, and metrics on production and staging.

+++ core/bin/ci/203_acceptance_tests.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+
+[ -z $SUPPRESS_TESTING ] || exit 0
 echo "Running acceptance tests"
 
 rm -rf tmp/capybara/
